+++ EngineSIU/EngineSIU/Saved/AutoSaves.scene
@@ -1,5 +1,35 @@
 {
-    "Actors": [],
+    "Actors": [
+        {
+            "ActorClass": "ACube",
+            "ActorID": "ACube_166",
+            "ActorLabel": "OBJ_CUBE_166",
+            "ActorTickInEditor": "false",
+            "Components": [
+                {
+                    "ComponentClass": "UStaticMeshComponent",
+                    "ComponentID": "StaticMeshComponent_0",
+                    "Properties": {
+                        "AABB_max": "X=0.410 Y=-0.000 Z=0.462",
+                        "AABB_min": "X=-0.409 Y=-1.899 Z=-0.462",
+                        "AttachParentID": "nullptr",
+                        "ComponentClass": "UStaticMeshComponent",
+                        "ComponentName": "StaticMeshComponent_0",
+                        "ComponentOwner": "ACube_166",
+                        "ComponentOwnerClass": "ACube",
+                        "RelativeLocation": "X=0.000 Y=0.000 Z=0.000",
+                        "RelativeRotation": "Pitch=0.000 Yaw=0.000 Roll=-0.000",
+                        "RelativeScale3D": "X=1.000 Y=1.000 Z=1.000",
+                        "StaticMeshPath": "Contents/base.obj",
+                        "bAutoActive": "true",
+                        "bIsActive": "true",
+                        "m_Type": ""
+                    }
+                }
+            ],
+            "RootComponentID": "StaticMeshComponent_0"
+        }
+    ],
     "NextUUID": 0,
     "Version": 1
 }
