+++ EngineSIU/EngineSIU/editor.ini
@@ -1,6 +1,6 @@
 WindowX=-8
 SplitterH.SplitRatio=0.500000
-PerspectiveCameraLocY0=-83.460152
+PerspectiveCameraLocY0=-0.275987
 PerspectiveCameraLocX3=0.000000
 PerspectiveCameraRotY1=0.000000
 SplitterV.SplitRatio=0.500000
@@ -9,15 +9,15 @@ ShowFlag1=63
 CameraSpeedSetting0=1
 CameraSpeedScalar0=3.265893
 PerspectiveCameraRotY2=0.000000
-PerspectiveCameraLocX0=-6.845141
+PerspectiveCameraLocX0=4.057710
 GridSize0=10.000000
-PerspectiveCameraLocZ0=147.155716
+PerspectiveCameraLocZ0=1.947681
 ShowFlag3=63
 CameraSpeedSetting2=1
 PerspectiveCameraRotX0=0.000000
 PerspectiveCameraLocX2=0.000000
-PerspectiveCameraRotY0=56.700478
-PerspectiveCameraRotZ0=-270.713165
+PerspectiveCameraRotY0=11.900030
+PerspectiveCameraRotZ0=-904.760925
 PerspectiveCameraRotX3=0.000000
 CameraSpeedSetting1=1
 ShowFlag0=2046
