+++ EngineSIU/EngineSIU/imgui.ini
@@ -15,8 +15,9 @@ Pos=1540,351
 Size=379,753
 
 [Window][Console]
-Pos=0,682
-Size=960,454
+Pos=0,-30
+Size=46,32
+Collapsed=1
 
 [Window][Engine Profiler]
 Pos=16,93
